Send a slack message on deletion (#7)
If you have SLACK_TOKEN and --slack-channel, c4 will send a message
to the selected Slack channel with a quick summary of which VMs have
been deleted.

I also chose to use regexes instead of just fixed strings. To use the regex
feature, use --aws-regex, --os-regex or --gcp-regex. By default, the regex
is `*.`.